job_application_api/requirements.txt
# Django and REST Framework
Django==5.0.0
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.1

# For environment variables
python-decouple==3.8

# For file/image uploads (resume files)
Pillow==10.1.0

# For better filtering and search
django-filter==23.5

